+++ src/main/java/de/cowtipper/cowlection/listener/skyblock/DungeonsListener.java
@@ -5,13 +5,13 @@ import de.cowtipper.cowlection.config.MooConfig;
 import de.cowtipper.cowlection.config.gui.MooConfigGui;
 import de.cowtipper.cowlection.data.DataHelper.DungeonClass;
 import de.cowtipper.cowlection.handler.DungeonCache;
+import de.cowtipper.cowlection.util.GuiHelper;
 import de.cowtipper.cowlection.util.TickDelay;
 import de.cowtipper.cowlection.util.Utils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.audio.SoundCategory;
 import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.inventory.GuiChest;
-import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.init.Blocks;
@@ -23,10 +23,6 @@ import net.minecraft.item.EnumDyeColor;
 import net.minecraft.item.ItemSkull;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.scoreboard.Score;
-import net.minecraft.scoreboard.ScoreObjective;
-import net.minecraft.scoreboard.ScorePlayerTeam;
-import net.minecraft.scoreboard.Scoreboard;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.MathHelper;
 import net.minecraftforge.client.event.ClientChatReceivedEvent;
@@ -34,11 +30,9 @@ import net.minecraftforge.client.event.GuiScreenEvent;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.common.util.Constants;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
-import net.minecraftforge.event.entity.player.PlayerSetSpawnEvent;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
-import net.minecraftforge.fml.relauncher.ReflectionHelper;
 import net.minecraftforge.fml.relauncher.Side;
 import org.apache.commons.lang3.StringUtils;
 import org.lwjgl.input.Mouse;
@@ -504,7 +498,7 @@ public class DungeonsListener {
             IInventory inventory = guiChest.inventorySlots.getSlot(0).inventory;
             if (inventory.getName().equals("Party Finder")) {
                 // get dungeon floor nr when joining a dungeon party via party finder
-                Slot hoveredSlot = getSlotUnderMouse(guiChest);
+                Slot hoveredSlot = GuiHelper.getSlotUnderMouse(guiChest);
                 if (hoveredSlot != null && hoveredSlot.getHasStack()) {
                     // clicked on an item
                     List<String> itemToolTip = hoveredSlot.getStack().getTooltip(Minecraft.getMinecraft().thePlayer, false);
@@ -516,7 +510,7 @@ public class DungeonsListener {
                 }
             } else if (inventory.getName().equals("Group Builder")) {
                 // get dungeon floor nr when creating a dungeon party for party finder
-                Slot hoveredSlot = getSlotUnderMouse(guiChest);
+                Slot hoveredSlot = GuiHelper.getSlotUnderMouse(guiChest);
                 if (hoveredSlot != null && hoveredSlot.getHasStack() && hoveredSlot.getStack().hasDisplayName()) {
                     // clicked on an item
                     String clickedItemName = EnumChatFormatting.getTextWithoutFormattingCodes(hoveredSlot.getStack().getDisplayName());
@@ -558,15 +552,6 @@ public class DungeonsListener {
                 main.getDungeonCache().setQueuedFloor(floorNr);
                 break;
             }
-        }
-    }
-
-    private Slot getSlotUnderMouse(GuiChest guiChest) {
-        try {
-            return ReflectionHelper.getPrivateValue(GuiContainer.class, guiChest, "theSlot", "field_147006_u");
-        } catch (ReflectionHelper.UnableToAccessFieldException e) {
-            e.printStackTrace();
-            return null;
         }
     }